Read events TSV

Wrapper around readr::read_tsv() to read in the raw events tab-delimited files. It provides the correct column type specification and disables the default quote escape option, as the files already have escaped quotes.

read_events_tsv(file, fix_names = TRUE, ...)

Arguments

file
Path to a raw events tab-delimited file (".tab").

fix_names
Normalize column names? Default is FALSE; see details. Names in the database are always normalized.

...
Other options passed to readr::read_tsv().

Details

The raw data file column names are capitalized and contain spaces. To make it easier to work with them in R and SQL, they are by default normalized by lower casing all and replacing spaces with underscores, e.g. "Event ID" becomes "event_id".