Get ICEWS Event Data

Download utilities and other helpers for getting the ICEWS event data from Dataverse.

Package options

Icews uses the following options() to configure behavior:

icews.debug: set to TRUE to print additional diagnostic messages in various functions.

Options that control file and database paths, and which should be set with setup_icews():

icews.data_dir: Path to the directory in which data will be kept locally. Depending on the other settings, subdirectories like "raw" and "db" will be created within.

icews.use_db: Use a SQLite database to store the data?

icews.keep_files: If "use_db" is TRUE, retain the raw data TSV files or delete after ingesting to the database?

See https://github.com/andybega/icews and https://www.andybeger.com/icews/ for more information.

References

Boschee, Elizabeth; Lautenschlager, Jennifer; O'Brien, Sean; Shellman, Steve; Starz, James; Ward, Michael, 2015, ``ICEWS Coded Event Data'', doi:10.7910/DVN/28075 , Harvard Dataverse.

Boschee, Elizabeth; Lautenschlager, Jennifer; O'Brien, Sean; Shellman, Steve; Starz, James, 2018, ``ICEWS Weekly Event Data'', doi:10.7910/DVN/QI2T9A , Harvard Dataverse.

See also

Author

Maintainer: Andreas Beger adbeger@gmail.com (ORCID)