Get DVN/local file/database state

Determine what data files, event sets, and version are currently on dataverse, in the local files, or in the local database.

get_dvn_manifest()

get_local_state(raw_file_dir = find_raw())

get_db_state(db_path = find_db())

Arguments

raw_file_dir
Directory containing raw data files

db_path
Path to SQLite database files.

Value

For get_local_state and get_db_state, a tibble with columns:

db/local_file: the full source data file name, e.g. "events.1995.20150313082510.tab".

db/local_data_set: the event set contained in the file, e.g. "events.1995".

db/local_version: the version of the file/event set, e.g. "20150313082510"

For get_dvn_manifest, a list of length 3, containing:

data_files: a tibble similar to those returned by get_local_state and get_db_state, with an additional column for the file label on dataverse since some files are zipped, i.e. ending with ".tab.zip" instead of ".tab".

files: a summary list of all files on DVN, consisting of the data files but also documentation and metadata files.

dataverse_dataset: an object of class "dataverse_dataset", returned by dataverse::get_dataset().

Details

The data files (tab-separated files, ".tab") on dataverse that contain the raw event data follow a common format denoting the set of events contained in a file and which version of the event data and/or file dump they correspond to. For example, "events.1995.20150313082510.tab" contains events for 1995 and the version is denoted by the timestamp, "20150313082510".

The download and update functions (update_icews(), download_data()) will recognize which event sets are locally available or still need to be downloaded, and whether any local even sets have been superseded by a new version in dataverse, by using

Examples

# NOT RUN {
# Remote (DVN) state
get_dvn_manifest()
# Local file state
get_local_state()
# Database state
get_db_state()
# }